Created an enum to hold all LocalStorage Keys; Created NewPostForm, FeedCard and FeedReaction components; Improved app flow about signout; Added axios interceptors;

+++ src/contexts/userContext/index.tsx
@@ -1,5 +1,6 @@
-import { createContext, Dispatch, useContext, useState } from 'react'
+import { createContext, Dispatch, useContext, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
+import { ELocalStorageKeys } from '../../@types'
 
 interface IUser_Context {
 	user: string | undefined
@@ -15,11 +16,18 @@ const UserProvider: React.FC = ({ children }) => {
 	const [user, setUser] = useState<string | undefined>(undefined)
 	const navigate = useNavigate()
 
-	const signOut = () => {
+	const signOut = () => setUser(undefined)
+
+	useEffect(() => {
+		if (user) return
+
 		localStorage.clear()
-		setUser(undefined)
 		navigate('/auth/sign-in')
-	}
+	}, [user])
+
+	useEffect(() => {
+		if (!localStorage.getItem(ELocalStorageKeys.ACESS_TOKEN)) setUser(undefined)
+	}, [localStorage.getItem(ELocalStorageKeys.ACESS_TOKEN)])
 
 	return (
 		<userContext.Provider value={{ user, setUser, signOut }}>
